Get code almost ready for release
