+++ README.md
@@ -1,3 +1,25 @@
-# Auto-channel
+# Comfy Channel
 
-Create a 24/7 live video broadcast with automatic content management and overlays
+Comfy Channel is a 24/7 live video broadcast with automatic content selection and overlays using FFMPEG and Python
+![Comfy Channel](screenshot.png)
+
+## How To
+
+The repository comes with some example content and configurations that can be used to set up Comfy Channel.
+
+More detailed instructions to come...
+
+## Inspiration
+
+When I'm working on a computer, I usually like to have live streams or videos running in the background or in a small window on the side. I don't really want to go out of my way to pick what I want to watch, so I decided to create my own streaming platform. I wanted to put a bunch of videos into it, and have it randomly select the content to play, showing what is up next once in a while just like some TV stations. However, my main goal for Comfy Channel was to have the little clock in the top left corner like you see on Japanese TV
+
+## Included Media
+
+Videos:
+[**Various videos that I recorded**](https://www.youtube.com/channel/UCjgBGlfTl5UMvtmH838x99w)
+
+Font:
+[**HC Maru Gothic font**](https://www.freejapanesefont.com/hc-maru-gothic-font-download/)
+
+Upnext audio files from:
+[**Delroy Edwards - SLOWED DOWN FUNK**](https://www.sloweddownfunk.net/)
