--- configs/nginx.conf
@@ -1,67 +1,0 @@
-worker_processes  auto;
-
-#error_log   /tmp/nginx_logs/error.log;
-
-events {
-    worker_connections  1024;
-}
-
-http {
-    include       mime.types;
-    default_type  application/octet-stream;
-
-    #log_format  main  '$remote_addr - $remote_user [$time_local] "$request" '
-    #                  '$status $body_bytes_sent "$http_referer" '
-    #                  '"$http_user_agent" "$http_x_forwarded_for"';
-
-    #access_log  /tmp/nginx_logs/access.log  main;
-
-    sendfile        on;
-    tcp_nopush     	on;
-    tcp_nodelay 	on;
-
-    keepalive_timeout  65;
-
-    server {
-        listen       80;
-        server_name  localhost;
-
-        location / {
-        	add_header 'Cache-Control' 'no-cache';
-            root html;
-            index index.html index.htm;
-
-            types {
-                application/dash+xml mpd;
-                application/vnd.apple.mpegurl m3u8;
-                video/mp2t ts;
-            }
-        }
-
-        location /stat {
-            rtmp_stat all;
-        }
-    }
-
-}
-
-rtmp {
-    server {
-        listen 1935;
-        ping 30s;
-        notify_method get;
-
-        application live{
-            live on;
-
-            # HLS Settings
-            hls on;
-            hls_type live;
-            hls_path hls;
-            hls_sync 200ms;
-            hls_fragment 1s;
-            hls_playlist_length 10s;
-        }
-
-    }
-}
