--- configs/playout.ini
@@ -1,23 +1,0 @@
-[Block1]
-name = "Test Block 1"
-folder = videos
-files = 2
-bump_chance = 0.2
-
-[Block2]
-name = "Test Block 2"
-folder = videos
-files = 3
-bump_chance = 0.2
-
-[Block3]
-name = "Test Block 3"
-folder = videos
-files = 4
-bump_chance = 0.2
-
-[Block4]
-name = "Test Block 4" 
-folder = videos
-files = 4
-bump_chance = 0.2
