+++ playout.ini
@@ -1,0 +1,11 @@
+[Block1]
+name = "Test Block 1"
+folder = videos
+files = 2
+bump_chance = 0.2
+
+[Block2]
+name = "Test Block 2"
+folder = videos
+files = 3
+bump_chance = 0.2
