--- requirements.txt
@@ -1,5 +1,0 @@
-colorama==0.4.1
-ffmpeg-python==0.1.17
-psutil==5.6.2
-pymediainfo==4.0
-
