+++ src/requirements.txt
@@ -1,0 +1,5 @@
+colorama
+ffmpeg-python
+psutil
+pymediainfo
+
