+++ upnext/wisdom.txt
@@ -1,0 +1,16 @@
+can I get a uhhhhhhhh....
+stay comfy
+hack the planet
+trust your technolust
+drink some water
+electric heaters are great
+mess with the best, die like the rest
+pool on the roof must have a leak
+this is a payphone
+dont ask
+type cookie
+youre in the butter zone now, baby
+chemicals
+get a free subdomain
+bladeless fans are cool
+KOPTAN tape
